Create caesars-cipher.js
caesars-cipher.js implements Caesar cipher/shift cipher. In a shift cipher, each alphabet is replaced by another alphabet shifted by a fixed number of places. This example implements the ROT13 cipher where the values of the letters are shifted by 13 places. So A - N, B - O, C - P and so on. So the code takes ROT13 encoded string as input and returns a decoded string. All uppercase letters will be decoded. Lowercase and non-alphabetic characters will not be decoded and returned as received.